[AppointmentFilterWrapper] pratiti | Renamed wrapper component
